Split firefox history and downloads SQLite plugins into separate files #3135
